move fragcomp
